Recombine extract fragments and agglomerate into single task
When we have separate tasks for extracting and agglomerating, we run
into a problem because agglomeration starts as soon as the edges
collection is created even though there is nothing in it yet, and
therefore nothing to agglomerate. We tried to fix this by creating a
separate collection 'complete' after extracting fragments blockwise.
We then returned that rather than nodes in the extract fragments task to
no avail. Here, we recombine the extract/agglomerate scripts and merge
the tasks. Extraction and agglomeration now works, but evaluation
doesn't - seems to not be able to find the db name...